Update README.md
This distro is a link to a publically available Google Fusion Table titled, " ." that demonstrates geocoding and heatmapping unemployment stats from 2011 by county for the entire us. As of publish date (5-1-2013) full 2012 stats are not available.

Open and save to your Google Account.  This is a large file and takes a while to load along with maps and data.  Three basic concepts are demonstrated on the Google Maps API: geocoding, heatmapping and using sort to focus on specific locales (PA & OH) to demonstrate splicing large datasets and mapping focus changes from the larger table (cards).  Try to shift the focus to your state and change heatmapping styles to gradient, etc.. It's easiest to copy one of the existing models in the file and then use it to create a new sort and map.

Here is the link: https://www.google.com/fusiontables/DataSource?docid=1PF3n3cLDlQhHdlP8BFndkQ-RtIAYiRznk-E1--s